Redux store and reducers conversion to TypeScript.
