--- src/renderer/store.js
@@ -1,12 +1,0 @@
-import { createStore, compose } from 'redux';
-import reducers from '../redux/reducers/index';
-
-// Enhancers function uses compose function to pass the Redux Store to the Redux extension.
-const enhancers = compose(
-  window.__REDUX_DEVTOOLS_EXTENSION__
-  && window.__REDUX_DEVTOOLS_EXTENSION__()
-);
-
-const store = createStore(reducers, enhancers);
-
-export default store;
