+++ src/renderer/store.ts
@@ -1,0 +1,21 @@
+import { configureStore } from '@reduxjs/toolkit'
+import reducers from '../redux/reducers/index';
+import { composeWithDevTools } from '@redux-devtools/extension';
+
+const store = configureStore({
+  reducer: reducers,
+  // this is not best practice!!
+  middleware: getDefaultMiddleware =>
+    getDefaultMiddleware({
+      serializableCheck: false,
+    }),
+})
+
+export type RootState = ReturnType<typeof store.getState>
+export type AppDispatch = typeof store.dispatch
+export default store;
+
+
+
+
+
